Fix building of color_mask for many colors
When the number of colors is larger than a machine word, the current code to build color_mask can go awry.  Here is why we think that's so:

color_mask is an appropriate-sized bitset, but colors[edge * num_colors + c] is an sd_vector, and sd_vector::operator[]() returns a "value_type" which is, I think, an unsigned long or unsigned long long.  So, if the shift is longer than the number of bits in "value_type" the result is wrong. This only shows up, of course, if the number of colors is reasonably large (e.g., greater than the number of bits in a machine word).  However, this changed line should fix it.
		    //color_mask |= colors[edge * num_colors + c] << c;